+++ docs/documentation.jsx
@@ -20,9 +20,11 @@ import * as reactEdit from 'react-edit';
 import * as reactabular from 'reactabular';
 import * as dnd from 'reactabular-dnd';
 import * as tree from 'reactabular-tree';
+import * as select from 'reactabular-select';
 import VisibilityToggles from 'reactabular-visibility-toggles';
 import * as Virtualized from 'reactabular-virtualized';
 import * as stylesheet from 'stylesheet-helpers';
+import selectabular from 'selectabular';
 
 import * as rowsDefinitions from './data/definitions';
 import * as customHelpers from './helpers';
@@ -32,7 +34,6 @@ const documentationImports = {
   Table: { ...reactabular.Table },
   Sticky: { ...reactabular.Sticky },
   search: { ...reactabular.search },
-  select: { ...reactabular.select },
   sort: { ...reactabular.sort },
   edit: { ...reactEdit },
   highlight: { ...reactabular.highlight },
@@ -46,6 +47,8 @@ const documentationImports = {
   Virtualized: { ...Virtualized },
   EasyTable,
   stylesheet: { ...stylesheet },
+  select: { ...selectabular },
+  ...select,
   ...reactDnd,
   classnames,
   compose,