+++ packages/reactabular-select/src/index.jsx
@@ -1,2 +1,2 @@
-export { default as byArrowKeys } from './by-arrow-keys';
-export { default as row } from './row';
+export { default as byArrowKeys } from './by-arrow-keys'; // eslint-disable-line import/prefer-default-export
+